+++ socket_server/src/events/coords.ts
@@ -1,7 +1,7 @@
 import { Socket } from "socket.io";
 import { Coord, Animation } from "../game-types";
 
-export function coords(coords: Coord, animState: Animation, sok: Socket): void {
+export function coords(x: number, y: number, animState: Animation, sok: Socket): void {
    if (sok.color === null) {
       sok.emit('error', 'coords: You are a spectator.');
       return;
@@ -15,8 +15,6 @@ export function coords(coords: Coord, animState: Animation, sok: Socket): void {
       return;
    }
    
-   coords.x = Math.floor(coords.x);
-   coords.y = Math.floor(coords.y);
-   sok.coords = coords;
+   sok.coords = { x: Math.floor(x), y: Math.floor(y) };
    sok.animState = animState;
 }