+++ spright/pyromasters-multiplayer_config/run.sh
@@ -1,3 +1,2 @@
 #!/bin/bash
-cd $(dirname "$0")
-../bin/spright --input spright.conf --path ../../src/client/assets/images/animations
+../bin/spright --input spright.conf --path ../../web/assets/images/animations